+++ handlers/pouchDBPluginServiceHandler.go
@@ -172,6 +172,7 @@ func getPouchDBPluginServiceDatastore() (db.PouchDBPluginServiceDatastore, error
 // on the API format.
 func (psh *PouchDBPluginServiceHandler) GetChanges(w http.ResponseWriter, r *http.Request) {
 	startTime := time.Now()
+	mon.IncrementCounter(mon.TenantAPIRecieved)
 
 	dbName := getDBFieldFromRequest(r, 2)
 
@@ -182,7 +183,7 @@ func (psh *PouchDBPluginServiceHandler) GetChanges(w http.ResponseWriter, r *htt
 	result, err := psh.pouchPluginDB.GetChanges(dbName, &queryParams)
 	if err != nil {
 		http.Error(w, fmt.Sprintf("Unable to retrieve %s: %s", db.ChangeFeedStr, err.Error()), http.StatusBadRequest)
-		mon.TrackAPITimeMetricInSeconds(startTime, "400", getChangesStr)
+		trackAPIMetrics(mon.PouchAPICompleted, startTime, "400", getChangesStr)
 		return
 	}
 
@@ -191,11 +192,11 @@ func (psh *PouchDBPluginServiceHandler) GetChanges(w http.ResponseWriter, r *htt
 	response, err := json.Marshal(result)
 	if err != nil {
 		http.Error(w, fmt.Sprintf("Error generating %s response: %s", db.ChangeFeedStr, err.Error()), http.StatusBadRequest)
-		mon.TrackAPITimeMetricInSeconds(startTime, "400", getChangesStr)
+		trackAPIMetrics(mon.PouchAPICompleted, startTime, "400", getChangesStr)
 		return
 	}
 
-	mon.TrackAPITimeMetricInSeconds(startTime, "200", getChangesStr)
+	trackAPIMetrics(mon.PouchAPICompleted, startTime, "200", getChangesStr)
 	fmt.Fprintf(w, string(response))
 }
 
@@ -204,6 +205,7 @@ func (psh *PouchDBPluginServiceHandler) GetChanges(w http.ResponseWriter, r *htt
 // CouchDB documentation on this API.
 func (psh *PouchDBPluginServiceHandler) CheckAvailability(w http.ResponseWriter, r *http.Request) {
 	startTime := time.Now()
+	mon.IncrementCounter(mon.TenantAPIRecieved)
 
 	logger.Log.Info("Checking for CouchDB availability")
 
@@ -211,7 +213,7 @@ func (psh *PouchDBPluginServiceHandler) CheckAvailability(w http.ResponseWriter,
 	result, err := psh.pouchPluginDB.CheckAvailability()
 	if err != nil {
 		http.Error(w, fmt.Sprintf("Error checking CouchDB availability: %s", err.Error()), http.StatusInternalServerError)
-		mon.TrackAPITimeMetricInSeconds(startTime, "500", serverHBStr)
+		trackAPIMetrics(mon.PouchAPICompleted, startTime, "500", serverHBStr)
 		return
 	}
 
@@ -219,12 +221,12 @@ func (psh *PouchDBPluginServiceHandler) CheckAvailability(w http.ResponseWriter,
 	response, err := json.Marshal(result)
 	if err != nil {
 		http.Error(w, fmt.Sprintf("Error generating response: %s", err.Error()), http.StatusInternalServerError)
-		mon.TrackAPITimeMetricInSeconds(startTime, "500", serverHBStr)
+		trackAPIMetrics(mon.PouchAPICompleted, startTime, "500", serverHBStr)
 		return
 	}
 	logger.Log.Info("CouchDB server is available.\n")
 
-	mon.TrackAPITimeMetricInSeconds(startTime, "200", serverHBStr)
+	trackAPIMetrics(mon.PouchAPICompleted, startTime, "200", serverHBStr)
 	fmt.Fprintf(w, string(response))
 }
 
@@ -233,6 +235,7 @@ func (psh *PouchDBPluginServiceHandler) CheckAvailability(w http.ResponseWriter,
 // of CouchDB local documents.
 func (psh *PouchDBPluginServiceHandler) StoreDBSyncCheckpoint(w http.ResponseWriter, r *http.Request) {
 	startTime := time.Now()
+	mon.IncrementCounter(mon.TenantAPIRecieved)
 
 	dbName := getDBFieldFromRequest(r, 2)
 	logger.Log.Infof("Attempting to store %s to DB %s", db.DBSyncCheckpointStr, dbName)
@@ -242,14 +245,14 @@ func (psh *PouchDBPluginServiceHandler) StoreDBSyncCheckpoint(w http.ResponseWri
 	requestBody, err := getRequestBodyAsGenericObject(r)
 	if err != nil {
 		http.Error(w, fmt.Sprintf("Unable to read %s content: %s", db.DBSyncCheckpointStr, err.Error()), http.StatusBadRequest)
-		mon.TrackAPITimeMetricInSeconds(startTime, "500", storeCheckpointStr)
+		trackAPIMetrics(mon.PouchAPICompleted, startTime, "500", storeCheckpointStr)
 		return
 	}
 
 	result, err := psh.pouchPluginDB.StoreDBSyncCheckpoint(dbName, &queryParams, requestBody)
 	if err != nil {
 		http.Error(w, fmt.Sprintf("Unable to store %s: %s", db.DBSyncCheckpointStr, err.Error()), http.StatusInternalServerError)
-		mon.TrackAPITimeMetricInSeconds(startTime, "500", storeCheckpointStr)
+		trackAPIMetrics(mon.PouchAPICompleted, startTime, "500", storeCheckpointStr)
 		return
 	}
 
@@ -258,11 +261,11 @@ func (psh *PouchDBPluginServiceHandler) StoreDBSyncCheckpoint(w http.ResponseWri
 	response, err := json.Marshal(result)
 	if err != nil {
 		http.Error(w, fmt.Sprintf("Error generating %s response: %s", db.DBSyncCheckpointStr, err.Error()), http.StatusInternalServerError)
-		mon.TrackAPITimeMetricInSeconds(startTime, "500", storeCheckpointStr)
+		trackAPIMetrics(mon.PouchAPICompleted, startTime, "500", storeCheckpointStr)
 		return
 	}
 
-	mon.TrackAPITimeMetricInSeconds(startTime, "200", storeCheckpointStr)
+	trackAPIMetrics(mon.PouchAPICompleted, startTime, "200", storeCheckpointStr)
 	fmt.Fprintf(w, string(response))
 }
 
@@ -271,6 +274,7 @@ func (psh *PouchDBPluginServiceHandler) StoreDBSyncCheckpoint(w http.ResponseWri
 // of CouchDB local documents.
 func (psh *PouchDBPluginServiceHandler) GetDBSyncCheckpoint(w http.ResponseWriter, r *http.Request) {
 	startTime := time.Now()
+	mon.IncrementCounter(mon.TenantAPIRecieved)
 
 	dbName := getDBFieldFromRequest(r, 2)
 	dbMethod := getDBFieldFromRequest(r, 3)
@@ -287,11 +291,11 @@ func (psh *PouchDBPluginServiceHandler) GetDBSyncCheckpoint(w http.ResponseWrite
 	if err != nil {
 		if checkError(err, notFound) {
 			http.Error(w, fmt.Sprintf("%s %s does not exist", db.DBSyncCheckpointStr, documentID), http.StatusNotFound)
-			mon.TrackAPITimeMetricInSeconds(startTime, "404", getCheckpointStr)
+			trackAPIMetrics(mon.PouchAPICompleted, startTime, "404", getCheckpointStr)
 			return
 		}
 		http.Error(w, fmt.Sprintf("Unable to retrieve %s: %s", db.DBSyncCheckpointStr, err.Error()), http.StatusInternalServerError)
-		mon.TrackAPITimeMetricInSeconds(startTime, "500", getCheckpointStr)
+		trackAPIMetrics(mon.PouchAPICompleted, startTime, "500", getCheckpointStr)
 		return
 	}
 
@@ -300,11 +304,11 @@ func (psh *PouchDBPluginServiceHandler) GetDBSyncCheckpoint(w http.ResponseWrite
 	response, err := json.Marshal(result)
 	if err != nil {
 		http.Error(w, fmt.Sprintf("Error generating %s response: %s", db.DBSyncCheckpointStr, err.Error()), http.StatusInternalServerError)
-		mon.TrackAPITimeMetricInSeconds(startTime, "500", getCheckpointStr)
+		trackAPIMetrics(mon.PouchAPICompleted, startTime, "500", getCheckpointStr)
 		return
 	}
 
-	mon.TrackAPITimeMetricInSeconds(startTime, "200", getCheckpointStr)
+	trackAPIMetrics(mon.PouchAPICompleted, startTime, "200", getCheckpointStr)
 	fmt.Fprintf(w, string(response))
 }
 
@@ -314,6 +318,7 @@ func (psh *PouchDBPluginServiceHandler) GetDBSyncCheckpoint(w http.ResponseWrite
 // on the API.
 func (psh *PouchDBPluginServiceHandler) GetDBRevisionDiff(w http.ResponseWriter, r *http.Request) {
 	startTime := time.Now()
+	mon.IncrementCounter(mon.TenantAPIRecieved)
 
 	dbName := getDBFieldFromRequest(r, 2)
 
@@ -322,7 +327,7 @@ func (psh *PouchDBPluginServiceHandler) GetDBRevisionDiff(w http.ResponseWriter,
 	requestBody, err := getRequestBodyAsGenericObject(r)
 	if err != nil {
 		http.Error(w, fmt.Sprintf("Unable to read %s content: %s", db.DBRevDiffStr, err.Error()), http.StatusBadRequest)
-		mon.TrackAPITimeMetricInSeconds(startTime, "400", dbDiffStr)
+		trackAPIMetrics(mon.PouchAPICompleted, startTime, "400", dbDiffStr)
 		return
 	}
 
@@ -330,7 +335,7 @@ func (psh *PouchDBPluginServiceHandler) GetDBRevisionDiff(w http.ResponseWriter,
 	result, err := psh.pouchPluginDB.GetDBRevisionDiff(dbName, requestBody)
 	if err != nil {
 		http.Error(w, fmt.Sprintf("Unable to retrieve %s: %s", db.DBRevDiffStr, err.Error()), http.StatusBadRequest)
-		mon.TrackAPITimeMetricInSeconds(startTime, "400", dbDiffStr)
+		trackAPIMetrics(mon.PouchAPICompleted, startTime, "400", dbDiffStr)
 		return
 	}
 
@@ -339,11 +344,11 @@ func (psh *PouchDBPluginServiceHandler) GetDBRevisionDiff(w http.ResponseWriter,
 	response, err := json.Marshal(result)
 	if err != nil {
 		http.Error(w, fmt.Sprintf("Error generating %s response: %s", db.DBRevDiffStr, err.Error()), http.StatusInternalServerError)
-		mon.TrackAPITimeMetricInSeconds(startTime, "500", dbDiffStr)
+		trackAPIMetrics(mon.PouchAPICompleted, startTime, "500", dbDiffStr)
 		return
 	}
 
-	mon.TrackAPITimeMetricInSeconds(startTime, "200", dbDiffStr)
+	trackAPIMetrics(mon.PouchAPICompleted, startTime, "200", dbDiffStr)
 	fmt.Fprintf(w, string(response))
 }
 
@@ -352,6 +357,7 @@ func (psh *PouchDBPluginServiceHandler) GetDBRevisionDiff(w http.ResponseWriter,
 // CouchDB documentation of the API.
 func (psh *PouchDBPluginServiceHandler) BulkDBUpdate(w http.ResponseWriter, r *http.Request) {
 	startTime := time.Now()
+	mon.IncrementCounter(mon.TenantAPIRecieved)
 
 	dbName := getDBFieldFromRequest(r, 2)
 
@@ -368,7 +374,7 @@ func (psh *PouchDBPluginServiceHandler) BulkDBUpdate(w http.ResponseWriter, r *h
 	result, err := psh.pouchPluginDB.BulkDBUpdate(dbName, requestBody)
 	if err != nil {
 		http.Error(w, fmt.Sprintf("Unable to complete %s: %s", db.DBBulkUpdateStr, err.Error()), http.StatusBadRequest)
-		mon.TrackAPITimeMetricInSeconds(startTime, "400", bulkUpdateStr)
+		trackAPIMetrics(mon.PouchAPICompleted, startTime, "400", bulkUpdateStr)
 		return
 	}
 
@@ -377,17 +383,18 @@ func (psh *PouchDBPluginServiceHandler) BulkDBUpdate(w http.ResponseWriter, r *h
 	response, err := json.Marshal(result)
 	if err != nil {
 		http.Error(w, fmt.Sprintf("Error generating %s response: %s", db.DBBulkUpdateStr, err.Error()), http.StatusInternalServerError)
-		mon.TrackAPITimeMetricInSeconds(startTime, "500", bulkUpdateStr)
+		trackAPIMetrics(mon.PouchAPICompleted, startTime, "500", bulkUpdateStr)
 		return
 	}
 
-	mon.TrackAPITimeMetricInSeconds(startTime, "200", bulkUpdateStr)
+	trackAPIMetrics(mon.PouchAPICompleted, startTime, "200", bulkUpdateStr)
 	fmt.Fprintf(w, string(response))
 }
 
 // CheckDBAvailability - heartbeat for the given database.
 func (psh *PouchDBPluginServiceHandler) CheckDBAvailability(w http.ResponseWriter, r *http.Request) {
 	startTime := time.Now()
+	mon.IncrementCounter(mon.TenantAPIRecieved)
 
 	dbName := getDBFieldFromRequest(r, 2)
 	logger.Log.Infof("Checking for availability of DB %s", dbName)
@@ -397,11 +404,11 @@ func (psh *PouchDBPluginServiceHandler) CheckDBAvailability(w http.ResponseWrite
 	if err != nil {
 		if checkError(err, notFound) {
 			http.Error(w, fmt.Sprintf("DB %s does not exist", dbName), http.StatusNotFound)
-			mon.TrackAPITimeMetricInSeconds(startTime, "404", dbHBStr)
+			trackAPIMetrics(mon.PouchAPICompleted, startTime, "404", dbHBStr)
 			return
 		}
 		http.Error(w, fmt.Sprintf("Error checking availability of DB %s: %s", dbName, err.Error()), http.StatusInternalServerError)
-		mon.TrackAPITimeMetricInSeconds(startTime, "500", dbHBStr)
+		trackAPIMetrics(mon.PouchAPICompleted, startTime, "500", dbHBStr)
 		return
 	}
 
@@ -409,12 +416,12 @@ func (psh *PouchDBPluginServiceHandler) CheckDBAvailability(w http.ResponseWrite
 	response, err := json.Marshal(result)
 	if err != nil {
 		http.Error(w, fmt.Sprintf("Error generating response: %s", err.Error()), http.StatusInternalServerError)
-		mon.TrackAPITimeMetricInSeconds(startTime, "500", dbHBStr)
+		trackAPIMetrics(mon.PouchAPICompleted, startTime, "500", dbHBStr)
 		return
 	}
 	logger.Log.Infof("DB %s is available.\n", dbName)
 
-	mon.TrackAPITimeMetricInSeconds(startTime, "200", dbHBStr)
+	trackAPIMetrics(mon.PouchAPICompleted, startTime, "200", dbHBStr)
 	fmt.Fprintf(w, string(response))
 }
 
@@ -423,6 +430,7 @@ func (psh *PouchDBPluginServiceHandler) CheckDBAvailability(w http.ResponseWrite
 // Couch documentation of this API.
 func (psh *PouchDBPluginServiceHandler) GetAllDBDocs(w http.ResponseWriter, r *http.Request) {
 	startTime := time.Now()
+	mon.IncrementCounter(mon.TenantAPIRecieved)
 
 	dbName := getDBFieldFromRequest(r, 2)
 
@@ -431,7 +439,7 @@ func (psh *PouchDBPluginServiceHandler) GetAllDBDocs(w http.ResponseWriter, r *h
 	requestBody, err := getRequestBodyAsGenericObject(r)
 	if err != nil {
 		http.Error(w, fmt.Sprintf("Unable to read %s request content: %s", db.DBAllDocsStr, err.Error()), http.StatusBadRequest)
-		mon.TrackAPITimeMetricInSeconds(startTime, "400", allDocsStr)
+		trackAPIMetrics(mon.PouchAPICompleted, startTime, "400", allDocsStr)
 		return
 	}
 
@@ -439,7 +447,7 @@ func (psh *PouchDBPluginServiceHandler) GetAllDBDocs(w http.ResponseWriter, r *h
 	result, err := psh.pouchPluginDB.GetAllDBDocs(dbName, requestBody)
 	if err != nil {
 		http.Error(w, fmt.Sprintf("Unable to fetch %s: %s", db.DBAllDocsStr, err.Error()), http.StatusBadRequest)
-		mon.TrackAPITimeMetricInSeconds(startTime, "400", allDocsStr)
+		trackAPIMetrics(mon.PouchAPICompleted, startTime, "400", allDocsStr)
 		return
 	}
 
@@ -448,17 +456,18 @@ func (psh *PouchDBPluginServiceHandler) GetAllDBDocs(w http.ResponseWriter, r *h
 	response, err := json.Marshal(result)
 	if err != nil {
 		http.Error(w, fmt.Sprintf("Error generating %s response: %s", db.DBAllDocsStr, err.Error()), http.StatusInternalServerError)
-		mon.TrackAPITimeMetricInSeconds(startTime, "500", allDocsStr)
+		trackAPIMetrics(mon.PouchAPICompleted, startTime, "500", allDocsStr)
 		return
 	}
 
-	mon.TrackAPITimeMetricInSeconds(startTime, "200", allDocsStr)
+	trackAPIMetrics(mon.PouchAPICompleted, startTime, "200", allDocsStr)
 	fmt.Fprintf(w, string(response))
 }
 
 // CreateDB - provides the ability for pouch to create a couchDB.
 func (psh *PouchDBPluginServiceHandler) CreateDB(w http.ResponseWriter, r *http.Request) {
 	startTime := time.Now()
+	mon.IncrementCounter(mon.TenantAPIRecieved)
 
 	dbName := getDBFieldFromRequest(r, 2)
 
@@ -468,7 +477,7 @@ func (psh *PouchDBPluginServiceHandler) CreateDB(w http.ResponseWriter, r *http.
 	result, err := psh.AddDB(dbName)
 	if err != nil {
 		http.Error(w, fmt.Sprintf("Unable to create DB %s: %s", dbName, err.Error()), http.StatusBadRequest)
-		mon.TrackAPITimeMetricInSeconds(startTime, "400", createDbStr)
+		trackAPIMetrics(mon.PouchAPICompleted, startTime, "400", createDbStr)
 		return
 	}
 
@@ -477,11 +486,11 @@ func (psh *PouchDBPluginServiceHandler) CreateDB(w http.ResponseWriter, r *http.
 	response, err := json.Marshal(result)
 	if err != nil {
 		http.Error(w, fmt.Sprintf("Error generating DB creation response: %s", err.Error()), http.StatusInternalServerError)
-		mon.TrackAPITimeMetricInSeconds(startTime, "500", createDbStr)
+		trackAPIMetrics(mon.PouchAPICompleted, startTime, "500", createDbStr)
 		return
 	}
 
-	mon.TrackAPITimeMetricInSeconds(startTime, "200", createDbStr)
+	trackAPIMetrics(mon.PouchAPICompleted, startTime, "200", createDbStr)
 	fmt.Fprintf(w, string(response))
 }
 
@@ -489,6 +498,7 @@ func (psh *PouchDBPluginServiceHandler) CreateDB(w http.ResponseWriter, r *http.
 // See http://docs.couchdb.org/en/2.1.1/api/document/common.html for documentation of the API
 func (psh *PouchDBPluginServiceHandler) GetDBDoc(w http.ResponseWriter, r *http.Request) {
 	startTime := time.Now()
+	mon.IncrementCounter(mon.TenantAPIRecieved)
 
 	dbName := getDBFieldFromRequest(r, 2)
 	docID := getDBFieldFromRequest(r, 3)
@@ -500,7 +510,7 @@ func (psh *PouchDBPluginServiceHandler) GetDBDoc(w http.ResponseWriter, r *http.
 	result, err := psh.pouchPluginDB.GetDoc(dbName, docID, &queryParams, &r.Header)
 	if err != nil {
 		http.Error(w, fmt.Sprintf("Unable to retrieve %s %s: %s", db.DBDocStr, docID, err.Error()), http.StatusBadRequest)
-		mon.TrackAPITimeMetricInSeconds(startTime, "400", getDBDocStr)
+		trackAPIMetrics(mon.PouchAPICompleted, startTime, "400", getDBDocStr)
 		return
 	}
 
@@ -509,11 +519,11 @@ func (psh *PouchDBPluginServiceHandler) GetDBDoc(w http.ResponseWriter, r *http.
 	response, err := json.Marshal(result["data"]) // Only need the data portion of this wrapper object
 	if err != nil {
 		http.Error(w, fmt.Sprintf("Error generating %s response: %s", db.DBDocStr, err.Error()), http.StatusBadRequest)
-		mon.TrackAPITimeMetricInSeconds(startTime, "400", getDBDocStr)
+		trackAPIMetrics(mon.PouchAPICompleted, startTime, "400", getDBDocStr)
 		return
 	}
 
-	mon.TrackAPITimeMetricInSeconds(startTime, "200", getDBDocStr)
+	trackAPIMetrics(mon.PouchAPICompleted, startTime, "200", getDBDocStr)
 	fmt.Fprintf(w, string(response))
 }
 
@@ -521,6 +531,7 @@ func (psh *PouchDBPluginServiceHandler) GetDBDoc(w http.ResponseWriter, r *http.
 // There is no CouchDB documentation of the API.
 func (psh *PouchDBPluginServiceHandler) BulkDBGet(w http.ResponseWriter, r *http.Request) {
 	startTime := time.Now()
+	mon.IncrementCounter(mon.TenantAPIRecieved)
 
 	dbName := getDBFieldFromRequest(r, 2)
 
@@ -530,7 +541,7 @@ func (psh *PouchDBPluginServiceHandler) BulkDBGet(w http.ResponseWriter, r *http
 	requestBody, err := getRequestBodyAsGenericObject(r)
 	if err != nil {
 		http.Error(w, fmt.Sprintf("Unable to read %s content: %s", db.DBBulkGetStr, err.Error()), http.StatusBadRequest)
-		mon.TrackAPITimeMetricInSeconds(startTime, "400", bulkGetStr)
+		trackAPIMetrics(mon.PouchAPICompleted, startTime, "400", bulkGetStr)
 		return
 	}
 
@@ -538,7 +549,7 @@ func (psh *PouchDBPluginServiceHandler) BulkDBGet(w http.ResponseWriter, r *http
 	result, err := psh.pouchPluginDB.BulkDBGet(dbName, &queryParams, requestBody)
 	if err != nil {
 		http.Error(w, fmt.Sprintf("Unable to complete %s: %s", db.DBBulkGetStr, err.Error()), http.StatusBadRequest)
-		mon.TrackAPITimeMetricInSeconds(startTime, "400", bulkGetStr)
+		trackAPIMetrics(mon.PouchAPICompleted, startTime, "400", bulkGetStr)
 		return
 	}
 
@@ -547,11 +558,11 @@ func (psh *PouchDBPluginServiceHandler) BulkDBGet(w http.ResponseWriter, r *http
 	response, err := json.Marshal(result)
 	if err != nil {
 		http.Error(w, fmt.Sprintf("Error generating %s response: %s", db.DBBulkGetStr, err.Error()), http.StatusInternalServerError)
-		mon.TrackAPITimeMetricInSeconds(startTime, "500", bulkGetStr)
+		trackAPIMetrics(mon.PouchAPICompleted, startTime, "500", bulkGetStr)
 		return
 	}
 
-	mon.TrackAPITimeMetricInSeconds(startTime, "200", bulkGetStr)
+	trackAPIMetrics(mon.PouchAPICompleted, startTime, "200", bulkGetStr)
 	fmt.Fprintf(w, string(response))
 }